Experiment with separate `hyped_io` crates for individual STM32 boards
