+++ lib/io/Cargo.toml
@@ -4,6 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-embassy-stm32 = { version = "0.1.0", features = ["stm32l476rg"] }
 heapless = "0.8.0"
-
