+++ lib/io/stm32l476rg/src/lib.rs
@@ -1,0 +1,3 @@
+#![no_std]
+
+pub mod gpio;
